Share slot iteration between session adapters
`expandUpcomingSessions` (parent/gamer) and `expandAssignedSessionsToCards`
(gedu) each had a near-identical `enumerateOccurrences` — the
in-progress carryover, the start_date-pinned `futureCursorBase`, the
per-slot forward walk, and the merge-sort-and-trim. The two
implementations were copy-pasted, each with subtly different doc
prose; that's exactly where DST and start_date invariants drift over
time.

Move the loop into `session-occurrence.ts` as `enumerateRowOccurrences`
and export `OPEN_ENDED_OCCURRENCE_CAP` alongside it. Both adapters now
own only their row → card shaping. Existing adapter tests (29 across
both files) keep covering the iteration behaviors end-to-end, so the
helper inherits that coverage without a duplicate test file.
