+++ src/lib/session-occurrence.ts
@@ -114,3 +114,100 @@ export function endDateToCutoff(
   if (endDate === null) return null;
   return fromZonedTime(`${endDate}T23:59:59.999`, timezone);
 }
+
+/**
+ * Default cap for open-ended products (no `end_date`) — the parent/gamer
+ * Sessions list and the gedu My Groups list both surface the next N
+ * occurrences across all of a product's weekly slots. End-dated products
+ * (camps, events, dated runs) ignore the cap and emit every occurrence
+ * up to and including `end_date`.
+ *
+ * Shared so the gedu's view of a club shows the same horizon as the
+ * parent's.
+ */
+export const OPEN_ENDED_OCCURRENCE_CAP = 8;
+
+/**
+ * Expand one row's slots into the concrete (start, end) pairs the
+ * dashboards consume. Shared between `expandUpcomingSessions`
+ * (parent/gamer) and `expandAssignedSessionsToCards` (gedu) — the only
+ * difference between the callers was the per-row card shaping, not the
+ * iteration.
+ *
+ * Per-slot it first surfaces a still-in-its-window previous occurrence
+ * (the bit `getNextSessionStart` hides because it only ever returns
+ * *future* starts), then walks forward emitting future occurrences
+ * until either `endBoundary` is crossed or `cap` is reached. Per-slot
+ * lists are merged, sorted ascending, and trimmed to `cap` so the
+ * soonest N across all slots win.
+ *
+ * `cap = Infinity` means "no cap" — used for end-dated products where
+ * the date range is the natural bound.
+ */
+export function enumerateRowOccurrences(args: {
+  slots: SlotShape[];
+  timezone: string;
+  now: Date;
+  startBoundary: Date | null;
+  endBoundary: Date | null;
+  cap: number;
+  windowCloseMs: number;
+}): Array<{ start: Date; end: Date }> {
+  const { slots, timezone, now, startBoundary, endBoundary, cap, windowCloseMs } =
+    args;
+  const out: Array<{ start: Date; end: Date }> = [];
+  const perSlotCap = Number.isFinite(cap) ? cap : Number.POSITIVE_INFINITY;
+
+  // If the product hasn't started yet (start_date is in the future), pin
+  // the future-iteration cursor to "just before start_date" so the
+  // search lands on or after start_date. Otherwise iterate from real
+  // `now`. The prev-week-in-window check returns null in the
+  // not-started-yet case on its own (no occurrence can be in progress
+  // before the product opens), so no separate guard is needed here.
+  const futureCursorBase =
+    startBoundary !== null && startBoundary.getTime() > now.getTime()
+      ? new Date(startBoundary.getTime() - 1)
+      : now;
+
+  for (const slot of slots) {
+    const schedule = {
+      dayOfWeek: slot.weekday,
+      startTime: slot.startTime,
+      timezone,
+    };
+    const durationMs = slot.durationMinutes * 60_000;
+
+    let emitted = 0;
+
+    const inProgress = getCurrentInProgressOccurrence({
+      slot,
+      timezone,
+      now,
+      startBoundary,
+      endBoundary,
+      windowCloseMs,
+    });
+    if (inProgress) {
+      out.push(inProgress);
+      emitted += 1;
+    }
+
+    let cursor = futureCursorBase;
+    while (emitted < perSlotCap) {
+      const start = getNextSessionStart(schedule, { now: cursor });
+      if (endBoundary !== null && start.getTime() > endBoundary.getTime()) break;
+      const end = new Date(start.getTime() + durationMs);
+      out.push({ start, end });
+      // Step past this start so the next iteration finds the following
+      // occurrence rather than re-emitting this one. `getNextSessionStart`
+      // is required to return a strictly future date relative to `now`;
+      // see its TZ regression test — if that contract slips, this loop
+      // never advances and the renderer pegs.
+      cursor = new Date(start.getTime() + 60_000);
+      emitted += 1;
+    }
+  }
+
+  out.sort((a, b) => a.start.getTime() - b.start.getTime());
+  return Number.isFinite(cap) ? out.slice(0, cap) : out;
+}
